requirements.txt
streamlit
streamlit-pdf-viewer
pdfplumber
pdf2image
weasyprint
langchain-groq
langchain
firecrawl-py